% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arg-max.R
\name{arg_max}
\alias{arg_max}
\title{ArgMax}
\usage{
arg_max(FUN, inputs)
}
\arguments{
\item{FUN}{a function over which to calculate the argmax value. This can be a named function, an anonymous function, or can use the formula `~ .x*2` form from library(purrr).}

\item{inputs}{a vector of values to pass to FUN}
}
\value{
a vector of type double
}
\description{
ArgMax
}
\details{
Determines which given input maximizes a given function.
}
\section{Future}{

- allow for multiple input formats, and return type based on supplied format
- promote purrr to recommended cf required, and implement with base as a fallback
}

\examples{
arg_max(FUN = function(x) cos(x)**2, inputs = c(-2:2))
arg_max(~ .x ** 2, seq(-2, 2, by = 0.5))
}
