% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_tax.R
\name{calc_tax}
\alias{calc_tax}
\title{Calculate US Tax Burden}
\usage{
calc_tax(amount, tax_table = NA, total = TRUE)
}
\arguments{
\item{amount}{A numeric vector of the salaries against which to calculate taxes owed.}

\item{tax_table}{A data.frame() with columns "lower", "upper", and "rate", corresponding to the lower and upper bounds of the income bracket and the rate at which that bracket is taxed.}

\item{total}{A boolean. Should the tax burden be returned as the total owed, or as a vector of amount owed by marginal tax bracket?}
}
\value{
A numeric vector of tax values, the same length as the input amount vector. If total=FALSE, this will be a list of marginal tax values instead.
}
\description{
Calculate US Tax Burden
}
\details{
Calculates the amount of money owed for a given salary under the marginal progressive US tax code. By default, the burden is calculated using tax tables from FY2019.
}
\examples{
calc_tax(60000)
calc_tax(c(36000, 85000, 115000))
calc_tax(60000, total = FALSE)
}
