% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starwars_name_converter.R
\name{convert_name}
\alias{convert_name}
\title{Convert to Star Wars Name (a meme)}
\usage{
convert_name(name, letter = NA, reverse = FALSE)
}
\arguments{
\item{name}{a character vector of length 1, in format "Firstname Lastname"}

\item{letter}{a vowel. If blank, a random vowel will be chosen.}

\item{reverse}{logical. Should the modified last name go first? If FALSE (default)
the modified first name will be first.}
}
\value{
a string of the memeified name in Star Wars meme format
}
\description{
Convert to Star Wars Name (a meme)
}
\details{
From a Twitter meme:
"To find your Star Wars name, remove the fourth letter
of your LAST name, delete all letters in your last name after the
fifth, add a random vowel to the end of it,
 then add that to your first name spelled backwards.
}
\examples{
convert_name("Mike Honcho")
}
